Move decision modals into MoneyReportHeaderModalsContext
